requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-Migrate
Flask-SQLAlchemy
python-dotenv
cryptography
pyotp